requirements.txt
apprise==0.8.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
Markdown==3.3.1
oauthlib==3.1.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
